+++ autonomous_agent_api/prisma/migrations/20240327040548_0/migration.sql
@@ -1,0 +1,15 @@
+-- Drop existing table if it exists
+DROP TABLE IF EXISTS "Agent";
+
+-- CreateTable
+CREATE TABLE "Agent" (
+    "id" TEXT NOT NULL,
+    "name" TEXT NOT NULL,
+    "action" TEXT[] DEFAULT ARRAY[]::TEXT[],
+    "triggers" TEXT[] DEFAULT ARRAY[]::TEXT[],
+    "deleted_at" TIMESTAMP(3),
+    "created_at" TIMESTAMP(3),
+    "updated_at" TIMESTAMP(3),
+
+    CONSTRAINT "Agent_pkey" PRIMARY KEY ("id")
+);