+++ autonomous_agent_api/CHANGELOG.md
@@ -6,3 +6,16 @@
  - Added sample test case for demo api
  - Added docker compose and image builder for backend.
  - Added read me guide for backend setup/ installation.
+
+# TITLE - AGENT CRUD API , DATE- 2024-03-27/28 
+ - Added Repository and Service folder 
+ - Created Agent prisma model
+ - Added agent_router 
+   - created crud Api endpoints 
+     POST api for agent creation
+     GET api for listing agents
+     GET api for getting agent properties.
+     PUT api to update agent properties.
+     DELETE api for soft deletion of agent.
+ -created dtos model for request and response inside model directory
+